Time Capsule/requirements.txt
Django==4.2
celery==5.2.7
redis==4.5.5
django-ckeditor==6.7.3
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
python-dotenv==1.1.1 